requirements.txt
matplotlib==3.8.2
numpy==1.24.4
pandas==2.2.1
Pillow==10.2.0
streamlit==1.32.2
streamlit_drawable_canvas==0.9.3
tensorflow==2.16.1
